requirements.txt
fastapi
uvicorn[standard]
python-telegram-bot==20.7
asyncpg
httpx
python-dotenv